+++ CHANGELOG.md
@@ -7,6 +7,9 @@ CHANGELOG
 **CHANGES**
 - Do not require `PlacementGroup/Enabled` to be set to `true` when passing an existing `PlacementGroup/Id`.
 
+**BUG FIXES**
+- Fix the ability to export cluster's logs when using `export-cluster-logs` command with the `--filters` option.
+
 3.1.3
 ------